Roughed in the data structures
Oh man that was a lot. Created a Hash Table with a linked list for the
overflow buckets. Each node in the hash table contains a string key and
a linked list of characters associated with that key. Functions added
for creation/destruction of all structures, inserting into the hash
table, handling collisions, deleting a key from the hash table.