Hash table generator built!
Surprisingly not too difficult. The generator reads a file containing
words separated by white space and puts those words in a list. Then
reads each word in the list two characters at a time (key), and searches the
word list for matches to those two characters (key). If it finds a match, it
adds an entry to the hash table with the key and a value of the
character after the key.