Tracked down a silly memory leak
Boy that took a lot of work. mhtlist_insert was leaking memory, the
problem was that since it insert pushed things on the back of the list,
it looked for the back of the list by looping through the list until
next->next was null, then set next to the item. It should have (and now
does) correctly set next->next to the item. What a pain.